# [risk=low][RW-14654] It's much faster to query BQ every time than cache in the RWB DB