--- a/api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsBatchUpdateService.java
+++ b/api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsBatchUpdateService.java
@@ -1,13 +1,18 @@
 package org.pmiops.workbench.initialcredits;
 
+import static org.pmiops.workbench.utils.LogFormatters.formatDurationPretty;
+
 import com.google.cloud.bigquery.FieldValueList;
 import com.google.cloud.bigquery.QueryJobConfiguration;
+import com.google.common.base.Stopwatch;
 import jakarta.inject.Provider;
+import java.time.Duration;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import org.pmiops.workbench.api.BigQueryService;
 import org.pmiops.workbench.config.WorkbenchConfig;
@@ -26,10 +31,13 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class InitialCreditsBatchUpdateService {
+  private static final Logger log =
+      Logger.getLogger(InitialCreditsBatchUpdateService.class.getName());
 
   private final BigQueryService bigQueryService;
   private final GoogleProjectPerCostDao googleProjectPerCostDao;
   private final InitialCreditsService initialCreditsService;
+  private final Provider<Stopwatch> stopwatchProvider;
   private final Provider<WorkbenchConfig> workbenchConfigProvider;
   private final UserDao userDao;
   private final WorkspaceDao workspaceDao;
@@ -39,6 +47,7 @@ public class InitialCreditsBatchUpdateService {
       BigQueryService bigQueryService,
       GoogleProjectPerCostDao googleProjectPerCostDao,
       InitialCreditsService initialCreditsService,
+      Provider<Stopwatch> stopwatchProvider,
       Provider<WorkbenchConfig> workbenchConfigProvider,
       UserDao userDao,
       WorkspaceDao workspaceDao) {
@@ -46,6 +55,7 @@ public class InitialCreditsBatchUpdateService {
     this.googleProjectPerCostDao = googleProjectPerCostDao;
     this.initialCreditsService = initialCreditsService;
     this.userDao = userDao;
+    this.stopwatchProvider = stopwatchProvider;
     this.workbenchConfigProvider = workbenchConfigProvider;
     this.workspaceDao = workspaceDao;
   }
@@ -69,12 +79,20 @@ public class InitialCreditsBatchUpdateService {
                 Collectors.toMap(
                     DbGoogleProjectPerCost::getGoogleProjectId, DbGoogleProjectPerCost::getCost));
 
+    Stopwatch stopwatch = stopwatchProvider.get().start();
     Map<String, Double> userWorkspaceBQCosts2 =
-        getWorkspaceCostsFromBQ().entrySet().stream()
+        getAllWorkspaceCostsFromBQ().entrySet().stream()
             .filter(entry -> googleProjectsForUserSet.contains(entry.getKey()))
             .collect(
                 Collectors.groupingBy(
                     Entry::getKey, Collectors.summingDouble(Map.Entry::getValue)));
+    Duration elapsed = stopwatch.stop().elapsed();
+    log.info(
+        String.format(
+            "checkInitialCreditsUsage: Filtered %d workspace cost entries from BigQuery in %s",
+            userWorkspaceBQCosts2.size(), formatDurationPretty(elapsed)));
+
+    // TODO compare the two
 
     Set<DbUser> dbUserSet =
         userIdList.stream().map(userDao::findUserByUserId).collect(Collectors.toSet());
@@ -82,7 +100,7 @@ public class InitialCreditsBatchUpdateService {
     initialCreditsService.checkInitialCreditsUsageForUsers(dbUserSet, userWorkspaceBQCosts);
   }
 
-  public Map<String, Double> getWorkspaceCostsFromBQ() {
+  public Map<String, Double> getAllWorkspaceCostsFromBQ() {
     final QueryJobConfiguration queryConfig =
         QueryJobConfiguration.newBuilder(
                 "SELECT id, SUM(cost) cost FROM `"