--- a/api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsBatchUpdateService.java
+++ b/api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsBatchUpdateService.java
@@ -5,7 +5,6 @@ import static org.pmiops.workbench.utils.LogFormatters.formatDurationPretty;
 import com.google.cloud.bigquery.FieldValueList;
 import com.google.cloud.bigquery.QueryJobConfiguration;
 import com.google.common.base.Stopwatch;
-import com.google.common.collect.Streams;
 import jakarta.inject.Provider;
 import java.time.Duration;
 import java.util.HashMap;
@@ -17,10 +16,8 @@ import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import org.pmiops.workbench.api.BigQueryService;
 import org.pmiops.workbench.config.WorkbenchConfig;
-import org.pmiops.workbench.db.dao.GoogleProjectPerCostDao;
 import org.pmiops.workbench.db.dao.UserDao;
 import org.pmiops.workbench.db.dao.WorkspaceDao;
-import org.pmiops.workbench.db.model.DbGoogleProjectPerCost;
 import org.pmiops.workbench.db.model.DbUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -36,7 +33,6 @@ public class InitialCreditsBatchUpdateService {
       Logger.getLogger(InitialCreditsBatchUpdateService.class.getName());
 
   private final BigQueryService bigQueryService;
-  private final GoogleProjectPerCostDao googleProjectPerCostDao;
   private final InitialCreditsService initialCreditsService;
   private final Provider<Stopwatch> stopwatchProvider;
   private final Provider<WorkbenchConfig> workbenchConfigProvider;
@@ -46,14 +42,12 @@ public class InitialCreditsBatchUpdateService {
   @Autowired
   public InitialCreditsBatchUpdateService(
       BigQueryService bigQueryService,
-      GoogleProjectPerCostDao googleProjectPerCostDao,
       InitialCreditsService initialCreditsService,
       Provider<Stopwatch> stopwatchProvider,
       Provider<WorkbenchConfig> workbenchConfigProvider,
       UserDao userDao,
       WorkspaceDao workspaceDao) {
     this.bigQueryService = bigQueryService;
-    this.googleProjectPerCostDao = googleProjectPerCostDao;
     this.initialCreditsService = initialCreditsService;
     this.userDao = userDao;
     this.stopwatchProvider = stopwatchProvider;
@@ -72,30 +66,16 @@ public class InitialCreditsBatchUpdateService {
     // Create Map Key: googleProject and value: cost
     Stopwatch stopwatch = stopwatchProvider.get().start();
     Map<String, Double> userWorkspaceBQCosts =
-        Streams.stream(googleProjectPerCostDao.findAllByGoogleProjectId(googleProjectsForUserSet))
-            .collect(
-                Collectors.toMap(
-                    DbGoogleProjectPerCost::getGoogleProjectId, DbGoogleProjectPerCost::getCost));
-    Duration elapsed = stopwatch.stop().elapsed();
-    log.info(
-        String.format(
-            "checkInitialCreditsUsage: Retrieved %d workspace cost entries from DB in %s",
-            userWorkspaceBQCosts.size(), formatDurationPretty(elapsed)));
-
-    stopwatch.reset().start();
-    Map<String, Double> userWorkspaceBQCosts2 =
         getAllWorkspaceCostsFromBQ().entrySet().stream()
             .filter(entry -> googleProjectsForUserSet.contains(entry.getKey()))
             .collect(
                 Collectors.groupingBy(
                     Entry::getKey, Collectors.summingDouble(Map.Entry::getValue)));
-    elapsed = stopwatch.stop().elapsed();
+    Duration elapsed = stopwatch.stop().elapsed();
     log.info(
         String.format(
             "checkInitialCreditsUsage: Filtered %d workspace cost entries from BigQuery in %s",
-            userWorkspaceBQCosts2.size(), formatDurationPretty(elapsed)));
-
-    // TODO compare the two
+            userWorkspaceBQCosts.size(), formatDurationPretty(elapsed)));
 
     Set<DbUser> dbUserSet =
         userIdList.stream().map(userDao::findUserByUserId).collect(Collectors.toSet());
@@ -103,7 +83,7 @@ public class InitialCreditsBatchUpdateService {
     initialCreditsService.checkInitialCreditsUsageForUsers(dbUserSet, userWorkspaceBQCosts);
   }
 
-  public Map<String, Double> getAllWorkspaceCostsFromBQ() {
+  private Map<String, Double> getAllWorkspaceCostsFromBQ() {
     final QueryJobConfiguration queryConfig =
         QueryJobConfiguration.newBuilder(
                 "SELECT id, SUM(cost) cost FROM `"