--- a/api/src/test/java/org/pmiops/workbench/api/OfflineBillingControllerTest.java
+++ b/api/src/test/java/org/pmiops/workbench/api/OfflineBillingControllerTest.java
@@ -59,6 +59,7 @@ public class OfflineBillingControllerTest {
   }
 
   private void mockInitialCreditCosts() {
-    when(mockInitialCreditsBatchUpdateService.getWorkspaceCostsFromBQ()).thenReturn(workspaceCosts);
+    when(mockInitialCreditsBatchUpdateService.getAllWorkspaceCostsFromBQ())
+        .thenReturn(workspaceCosts);
   }
 }