Change the behavior of the checkInitialCreditsUsage cron so that it is no longer in danger of timing out **before** dispatching usage checks to users, by removing the RWB DB interaction step.  This step is very slow and also not necessary, because the individual user-check tasks can query BigQuery themselves instead of the RWB DB.

Add timing logs to demonstrate this change in run time.

---

Logic before, pre-queue:
* query BigQuery for all workspaces (a few **seconds** in Prod and Test)
* clear the RWB DB cache of workspace costs (**1 minute** in Prod, **1:30** in Test)
* save the BigQuery data in the RWB DB cache (**4 min 30** in Prod, **5m** in Test)
* queue up batches of user IDs

Logic before, post-queue:
* query RWB DB for the relevant google projects for the batch of users
* query the workspace costs cache in the RWB DB
* do all the necessary processing of users and workspaces related to updating cost information 

---

Logic after, pre-queue:
* queue up batches of user IDs

Logic after, post-queue:
* query RWB DB for the relevant google projects for the batch of users
* query BigQuery for all workspaces
  * limiting to just the relevant google projects is an option for future improvement here, but it didn't seem to do much in my testing
  * after further investigation, this is likely to be **much more expensive** (due to cache misses) for little benefit
* do all the necessary processing of users and workspaces related to updating cost information 

---

From a Local run, DB insert vs BQ retrieval
```
checkInitialCreditsUsage: Retrieved 41310 workspace cost entries from BigQuery in 4s 229ms
checkInitialCreditsUsage: cleared googleproject_cost table in 22s 194ms
checkInitialCreditsUsage: Inserted all workspace costs to googleproject_cost table in 20s 515ms
```

DB retrieval vs BQ retrieval
```
checkInitialCreditsUsage: Retrieved 11 workspace cost entries from DB in 10ms
checkInitialCreditsUsage: Filtered 11 workspace cost entries from BigQuery in 2s 336ms
```


---
**PR checklist**

- [x] I have included an issue ID or "no ticket" in the PR title as outlined in [CONTRIBUTING.md](https://github.com/all-of-us/workbench/blob/main/.github/CONTRIBUTING.md).
- [x] I have included a risk tag of the form `[risk=no|low|moderate|severe]` in the PR title as outlined in [CONTRIBUTING.md](https://github.com/all-of-us/workbench/blob/main/.github/CONTRIBUTING.md).
- [ ] I have manually tested this change and my testing process is described above.
- [ ] This change includes appropriate automated tests, and I have documented any behavior that cannot be tested with code.
- [ ] I have added explanatory comments where the logic is not obvious.
- One or more of the following is true:
  - [ ] This change is intended to complete a JIRA story, so I have checked that all AC are met for that story.
  - [ ] This change fixes a bug, so I have ensured the steps to reproduce are in the Jira ticket or provided above.
  - [ ] This change impacts deployment safety (e.g. removing/altering APIs which are in use), so I have documented the impacts in the description.
  - [ ] This change includes a new feature flag, so I have created and linked new JIRA tickets to (a) turn on the feature flag and (b) remove it later.
  - [ ] This change modifies the UI, so I have taken screenshots or recordings of the new behavior and notified the PO and UX designer in [Slack](https://pmi-engteam.slack.com/archives/C02MWP2RN5P).
  - [ ] This change modifies API behavior, so I have run the relevant E2E tests locally because API changes are not covered by our PR checks.
  - [ ] None of the above apply to this change.