@Qi77Qi 2025-04-03
do we get charged for BQ queries? just wonder what the impact would be