@Qi77Qi 2025-04-03
Just thoughts
1. look GCP docs to understand BQ query cost
2. let the change sit in Test for a while and see if cost for our control plane project got much higher (not sure if this super telling since Test probably have much less traffic)