@jmthibault79 2025-04-03
After investigation, I believe that this change will result in no additional charges, because the identical queries will be cached. In more detail:

Looking at this query (for Prod) in BQ explorer, I see "This query will process 55.47 GB when run."
```
SELECT id, SUM(cost) cost FROM `all-of-us-rw-prod-bdata.RwProdBillingDataExport.rw_firecloud_view`
WHERE id IS NOT NULL
GROUP BY id ORDER BY cost desc;
```

By the [on-demand pricing model](https://cloud.google.com/bigquery/pricing#on_demand_pricing), we are charged $6.25 per TiB (with the first 1 TiB per month free).

Before: one query every 30 min -> 48 runs / day -> 2.6 TB/day -> $16/day -> $500/mo
After: executes for ~150 batches each time.  We might expect this to be 150x in cost, but these queries should be [cached](https://cloud.google.com/bigquery/docs/cached-results) and therefore without cost.  I was able to confirm that executions of this specific query hit the cache on subsequent runs.

Note: updating this query to filter on project with `IN()` would cause each query to be a cache **miss** (confirmed in BQ explorer) yet charge us the cost for the full dataset, so this would indeed be a 150x increase in cost!