@jmthibault79 2025-04-03
Actually, I expect the $500/mo "Before" cost to be an overestimate, because many of the every-30-min queries should be cache hits, assuming the underlying costs have not changed.